requirements.txt
flask>=3.1.0
flask-sqlalchemy>=3.1.0
flask-migrate>=4.1.0
flask-marshmallow>=1.2.0
flask-redis>=0.4.0
flask-cors>=5.0.0
flask-limiter>=3.8.0
marshmallow-sqlalchemy>=1.1.0
marshmallow>=3.23.0
celery[redis]>=5.4.0
alembic>=1.14.0
gunicorn>=22.0.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.1
redis>=5.2.0
